pyproject.toml
[project]
name = "fact-check-in"
version = "0.7.0"
description = "Validate if the attendee is registered for the conference by ticket code, name and email."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "mkdocs>=1.6.1",
    "mkdocs-macros-plugin>=1.3.7",
    "mkdocs-material[imaging]>=9.6.5",
    "omegaconf>=2.3.0",
    "pydantic[email]>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "starlette-context>=0.3.6",
    "structlog>=25.1.0",
    "unidecode>=1.3.8",
    "uvicorn>=0.34.0",
]

line-length = 140
indent-width = 4
target-version = ['py312']

[dependency-groups]
dev = [
    "faker>=36.1.1",
    "httpx>=0.28.1",
    "hypothesis>=6.126.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
]


[tool.ruff]
line-length = 140
target-version = "py312"
extend-exclude = ["*.md"]

[tool.ruff.lint]
select = [
    "F",  # pyflakes
    "E", "W",  # pycodestyle
    "A",  # flake8-builtins
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM",  # flake8-simplify
    "ARG",  # flake8-unused-arguments
    "PL",  # pylint
    "TID",  # tidy
    "I",  # isort
    "N",  # pep8-naming
    "UP"  # pyupgrade
]
ignore = [
    "E501",  # may cause conflicts with ruff formatter
    "W191"
]
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore:UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:Provider `faker.providers.*',
    'ignore:Looking for locale `en_US` in provider.*',
]
